+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 5.5.2
+🐛 **Bug Fixes**
+- fixed an issue where reordering stopped working after changing device orientation
+
+🏄‍♂️ **Improvements**
+- Updated not recommended `MediaQuery` calls:
+  - `MediaQuery.of(context).size` → `MediaQuery.sizeOf(context)`
+  - `MediaQuery.of(context).orientation` → `MediaQuery.orientationOf(context)`
+- Thanks to `hajajmaor` for reporting this! ([#163](https://github.com/karvulf/flutter-reorderable-grid-view/issues/163))
+
+
 ## 5.5.1
 🐛 **Bug Fixes**
 - forgotten assert in the `builder` constructor was added (thanks to `FaFre`, see [#159](https://github.com/karvulf/flutter-reorderable-grid-view/pull/159))